% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bigdata}
\alias{bigdata}
\title{Snippet of example data to demonstrate the functionality of kronos betwween and among three different groups}
\format{
A long format data.frame object with three 113 columns, features, and 94 rows, samples.
}
\usage{
bigdata
}
\description{
Snippet of example data to demonstrate the functionality of kronos betwween and among three different groups
}
\keyword{datasets}
