% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bigmeta}
\alias{bigmeta}
\title{Descriptional metadata for the 'bigdata' object, for the purpose of demonstration.}
\format{
A long format data.frame object with three 113 columns and 94 rows.
}
\source{
Dummy metadata modified from untargeted hippocampal metabolomics over the day
}
\usage{
bigmeta
}
\description{
Descriptional metadata for the 'bigdata' object, for the purpose of demonstration.
}
\keyword{datasets}
