% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronos_2.R
\name{build_kronos_formula}
\alias{build_kronos_formula}
\title{Update kronos formula in light of sine and cosine components}
\usage{
build_kronos_formula(formula, time, verbose)
}
\description{
Update fronos formula.
}
