% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{delistKronos_groupwise}
\alias{delistKronos_groupwise}
\title{Get Results from list of KronosOut Objects}
\usage{
delistKronos_groupwise(kronos_list, padjust = T)
}
\arguments{
\item{kronos_list}{a list of preferrably named kronosOut objects.}

\item{padjust}{a boolean. Toggles FDR using Benjamini Hochbergs procedure.}
}
\value{
A table with circadian output stats per group per feature.
}
\description{
These functions provides a unified wrapper to retrieve results
 from a list of \code{kronosOut} objects.
}
\examples{
\dontrun{
#Load prepared data stored in Kronos library
data("kronos_demo")
output <- kronos(formula = Variable_1 ~ time(Timepoint), 
data = onevariable, period = 24, verbose = T, pairwise = F)

#Extracting data from the output object:
getKronos_fit(output)
getKronos_trace(output)
getKronos_groupwise(output)

#Plotting:
gg_kronos_circle(output)
gg_kronos_sinusoid(output)

#For high-dimensional data, use fw_fronos:
out_list = fw_kronos(x = bigdata, formula = ~ Group + time(Timepoint), 
metadata = bigmeta, period = 24, verbose = F, pairwise = T) 

#Extracting data from the output object: 
kronosListToTable(out_list)


#Plotting:
gg_kronos_acrogram(out_list)
}
}
