% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{delistKronos_pairwise}
\alias{delistKronos_pairwise}
\title{Get Results from list of \code{KronosOut} Objects}
\usage{
delistKronos_pairwise(kronos_list, padjust = TRUE)
}
\arguments{
\item{kronos_list}{a list of preferrably named \code{kronosOut} objects.}

\item{padjust}{a boolean. Toggles FDR using Benjamini Hochbergs procedure.}
}
\value{
ANOVA-like adjusted p-values for how each factor interacts with time.
}
\description{
These functions provides a unified wrapper to retrieve results
 from a list of \code{kronosOut} objects.
}
