% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fit_cosinor_model}
\alias{fit_cosinor_model}
\title{Fit cosinor model}
\usage{
fit_cosinor_model(x, measurement, groups, time, verbose = T, for_pw = F)
}
\description{
Fit cosinor model for totality of data
}
