% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronos_2.R
\name{fit_groupwise_model_test}
\alias{fit_groupwise_model_test}
\title{Fit cosinor model}
\usage{
fit_groupwise_model_test(
  data,
  group,
  time = time,
  period = period,
  verbose = T
)
}
\description{
Fit cosinor model for one aspect of data
}
