% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronos.R
\name{fw_kronos}
\alias{fw_kronos}
\title{Fit a cosinor model and extract relevant parameters on a feature table.}
\usage{
fw_kronos(
  x,
  formula,
  metadata,
  time = NULL,
  period = 24,
  verbose = F,
  pairwise = F
)
}
\arguments{
\item{x}{Input data. A table with rows being features and columns being samples}

\item{formula}{A formula. Use the time() function to designate which variable represents time. Leave the left-hand side of the formula empty as it will be sequentially replaced by every feature in the table.}

\item{metadata}{A metadata table, with rows being samples and columns being metadata entries}

\item{time}{A string. Should be the column name containing the time values.}

\item{period}{A numeric. The length of a period, in the same format as the \code{time} parameter.}

\item{verbose}{A boolean. Toggles whether to print diagnostic information while running. Useful for debugging errors on large datasets.}

\item{pairwise}{A boolean. Toggles whether to perform pairwise ANOVAs as a TukeyHSD-like post-hoc.}
}
\value{
A list of kronosOut S4 objects containing coefficients and all operations.
}
\description{
This wrapper applies kronos(), the main workhorse function in the kronos package. It manages the individual functionalities of kronos, including rhythmicity analysis and differential rhythmicity.
}
