% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronos.R
\name{get_cos_sine}
\alias{get_cos_sine}
\title{Get sine and cosine components}
\usage{
get_cos_sine(data, period, colnamePrefix = NULL)
}
\arguments{
\item{data}{input data}

\item{period}{A numeric. The length of a period, in the same format as the \code{time} parameter.}

\item{colnamePrefix}{A character string. Typically the name of the "Time" variable.}
}
\description{
Based on 'cosinor' and 'limorhyde' packages
}
