% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronos_2.R
\name{get_response}
\alias{get_response}
\title{Figure out what variable represents the response variable. Called by main kronos function.}
\usage{
get_response(formula, verbose = T)
}
\arguments{
\item{formula}{A formula. Use the time() function to designate which variable represents time.}

\item{verbose}{A boolean. Toggles whether to print diagnostic information while running. Useful for debugging errors on large datasets.}
}
\description{
Extracts the response variable from the formula.
}
