% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_rhythmic_trace}
\alias{get_rhythmic_trace}
\title{Get coordinates for plotting rhythmic traces}
\usage{
get_rhythmic_trace(fit, groups, allgroups, time, period, verbose = T)
}
\description{
Calculates coordinates to plot traces for rhythmic data based on /code{fit_groupwise_model}
}
