% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronos_2.R
\name{get_vars}
\alias{get_vars}
\title{Figure out what variable represents time. Called by main kronos function.}
\usage{
get_vars(formula, time, data, verbose = T)
}
\arguments{
\item{formula}{A formula. Use the time() function to designate which variable represents time.}

\item{time}{A string. Should be the column name containing the time values.}

\item{data}{input data}

\item{verbose}{A boolean. Toggles whether to print diagnostic information while running. Useful for debugging errors on large datasets.}
}
\description{
Extracts time from the formula and from the time argument. Also handles inconsistencies.
}
