% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getKronos_fit}
\alias{getKronos_fit}
\title{Get Results from KronosOut Object}
\usage{
getKronos_fit(kronosOut)
}
\arguments{
\item{kronosOut}{a \code{kronosOut} output object.}
}
\value{
The model fit used.
}
\description{
These functions provides a unified wrapper to retrieve results
 from a \code{kronosOut} object.
}
