% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getKronos_groupwise}
\alias{getKronos_groupwise}
\title{Get Results from KronosOut Object}
\usage{
getKronos_groupwise(kronosOut)
}
\arguments{
\item{kronosOut}{a \code{kronosOut} output object.}
}
\value{
Rhythmicity parameters per group.
}
\description{
These functions provides a unified wrapper to retrieve results
 from a \code{kronosOut} object.
}
