% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getKronos_input}
\alias{getKronos_input}
\title{Get Results from KronosOut Object}
\usage{
getKronos_input(kronosOut)
}
\arguments{
\item{kronosOut}{a \code{kronosOut} output object.}
}
\value{
The data used as input for the model.
}
\description{
These functions provides a unified wrapper to retrieve results
 from a \code{kronosOut} object.
}
