% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getKronos_pairwise}
\alias{getKronos_pairwise}
\title{Get Results from KronosOut Object}
\usage{
getKronos_pairwise(kronosOut)
}
\arguments{
\item{kronosOut}{a \code{kronosOut} output object.}
}
\value{
Pairwise comparisons between groups.
}
\description{
These functions provides a unified wrapper to retrieve results
 from a \code{kronosOut} object.
}
