% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getKronos_pairwise_p}
\alias{getKronos_pairwise_p}
\title{Get Results from KronosOut Object}
\usage{
getKronos_pairwise_p(kronosOut)
}
\arguments{
\item{kronosOut}{a \code{kronosOut} output object.}
}
\value{
ANOVA-like adjusted p-values for how each factor interacts with time.
}
\description{
These functions provides a unified wrapper to retrieve results
 from a \code{kronosOut} object.
}
