% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getKronos_params}
\alias{getKronos_params}
\title{Get Results from KronosOut Object}
\usage{
getKronos_params(kronosOut)
}
\arguments{
\item{kronosOut}{a \code{kronosOut} output object.}
}
\value{
The names and values of additional circadian model parameters, mostly for plotting purposes.
}
\description{
These functions provides a unified wrapper to retrieve results
 from a \code{kronosOut} object.
}
