% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getKronos_trace}
\alias{getKronos_trace}
\title{Get Results from KronosOut Object}
\usage{
getKronos_trace(kronosOut)
}
\arguments{
\item{kronosOut}{a \code{kronosOut} output object.}
}
\value{
The traces per group for plotting.
}
\description{
These functions provides a unified wrapper to retrieve results
 from a \code{kronosOut} object.
}
