% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{gg_kronos_acrogram}
\alias{gg_kronos_acrogram}
\title{A plotting method for acrophase circleplots using ggplot.}
\usage{
gg_kronos_acrogram(kronosOutList)
}
\arguments{
\item{kronosOutList}{A list of KronosOut output files from the main kronos() function.}
}
\description{
Wrapper around ggplot to make circadian circleplots.
}
