% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronos_2.R
\name{kronos}
\alias{kronos}
\title{Fit a cosinor model and extract relevant parameters}
\usage{
kronos(formula, data, time = NULL, period = 24, verbose = T, pairwise = T)
}
\arguments{
\item{formula}{A formula. Use the time() function to designate which variable represents time.}

\item{data}{input data}

\item{time}{A string. Should be the column name containing the time values.}

\item{period}{A numeric. The length of a period, in the same format as the \code{time} parameter.}

\item{verbose}{A boolean. Toggles whether to print diagnostic information while running. Useful for debugging errors on large datasets.}

\item{pairwise}{A boolean. Toggles whether to perform pairwise ANOVAs as a TukeyHSD-like post-hoc.}
}
\value{
A kronosOut S4 object containing coefficients and all operations.
}
\description{
This is the main workhorse function in the kronos package. It manages the individual functionalities of kronos, including rhythmicity analysis and differential rhythmicity.
}
