% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronos.R
\name{kronos_anova}
\alias{kronos_anova}
\title{Extract p-value from full fit}
\usage{
kronos_anova(fit, time)
}
\arguments{
\item{fit}{A lm model fit.}

\item{time}{A string. Should be the column name containing the time values.}
}
\description{
Compute p-values from full fit.
}
