% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronos_2.R
\name{kronos_predict}
\alias{kronos_predict}
\title{Give tracing information for plotting purposes}
\usage{
kronos_predict(fit, period, time, verbose = T)
}
\arguments{
\item{fit}{A model fit. Can be found in KronosOut@fit}

\item{verbose}{A boolean. Toggles whether to print diagnostic information while running. Useful for debugging errors on large datasets.}
}
\description{
Generate data needed to plot cosinor trace line.
}
