% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{kronosListToTable}
\alias{kronosListToTable}
\title{Wrangle results from list of KronosOut Objects to publication ready table.}
\usage{
kronosListToTable(kronos_list, padjust = T)
}
\arguments{
\item{kronos_list}{a list of preferrably named kronosOut objects.}

\item{padjust}{a boolean. Toggles FDR using Benjamini Hochbergs procedure.}
}
\value{
A table with circadian output stats per group per feature.
}
\description{
These functions provides a unified wrapper to retrieve results
 from a list of \code{kronosOut} objects.
}
