% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{kronosOut-class}
\alias{kronosOut-class}
\title{An S4 class to contain all output parameters and plotting information from the 'kronos' main pipeline}
\description{
kronosOut is the main output container of the main 'kronos' functions.
}
\section{Slots}{

\describe{
\item{\code{input}}{A data.frame with the data that was fed to the main workhorse function as 'x'}

\item{\code{fit}}{An lm fit for the entire model for the purpose of assessing differential rhytmicity.}

\item{\code{to_plot}}{A data.frame with the traces required to plot individual sinusoid curves}

\item{\code{ind_fit}}{A data.frame with the parameters from individual rhythmic model fits.}

\item{\code{pairwise_t}}{A data.frame with the p.values for pairwise comparisons, if applicable.}
}}

