% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{onevariable}
\alias{onevariable}
\title{Snippet of example data to demonstrate the functionality of 'kronos' in the most simple scenario.}
\format{
A long format data.frame object with three columns, and 31 rows, samples.
}
\source{
Dummy data modified from metagenomic analysis of mouse caecal contents over the day to demonstrate rhythmicity within one group.
}
\usage{
onevariable
}
\description{
Snippet of example data to demonstrate the functionality of 'kronos' in the most simple scenario.
}
\keyword{datasets}
