% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pairwise_cosinor_model}
\alias{pairwise_cosinor_model}
\title{Fit pairwise cosinor models as some sort of TukeyHSD.}
\usage{
pairwise_cosinor_model(
  x = x,
  measurement = measurement,
  groups = groups,
  time = time,
  pairwise_t = pairwise_t,
  verbose = verbose
)
}
\description{
Fit cosinor model for subset of data.
}
