% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronos_2.R
\name{pairwise_cosinor_model_test}
\alias{pairwise_cosinor_model_test}
\title{Fit pairwise cosinor models as some sort of TukeyHSD.}
\usage{
pairwise_cosinor_model_test(data, formula, time, verbose)
}
\description{
Fit cosinor model for subset of data.
}
