% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pairwise_rhythm}
\alias{pairwise_rhythm}
\title{Fit a cosinor model and extract relevant parameters}
\usage{
pairwise_rhythm(x, measurement, groups, time, period = 24, verbose = T)
}
\arguments{
\item{x}{input data}

\item{measurement}{A string. Should be the column name containing the meansurement data}

\item{groups}{A string. Should be the column name containing the group names.}

\item{time}{A string. Should be the column name containing the time values.}

\item{period}{A numeric. The length of a period, in the same format as the \code{time} parameter.}

\item{verbose}{A boolean. Toggles whether to print diagnostic information while running. Useful for debugging errors on large datasets.}
}
\value{
A list containing coefficients and all operations.
}
\description{
This is the main workhorse function in the kronos package. It manages the individual functionalities of kronos, including rhythmicity analysis and differential rhythmicity.
}
