% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{show,kronosOut-method}
\alias{show,kronosOut-method}
\title{Show method for \code{kronosOut} object.}
\usage{
\S4method{show}{kronosOut}(object)
}
\arguments{
\item{object}{An \code{kronosOut} object.}
}
\value{
Does not return anything, for efficiency reasons. The obvious side effect is output to the terminal.
}
\description{
method to print \code{kronosOut} object by calling \code{show}.
Since kronosOut objects are typically unwieldy, also gives some tips on how to handle it.
}
