% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronos_2.R
\name{std_test}
\alias{std_test}
\title{Calculate standard error}
\usage{
std_test(x, na.rm = TRUE)
}
\arguments{
\item{x}{input vector}
}
\value{
standard error of the input vector
}
\description{
calculate standard error for plotting purposes
}
