% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twowaydata}
\alias{twowaydata}
\title{Snippet of example data to demonstrate the functionality of 'kronos' in the two-factor design scenario.}
\format{
A wide format data.frame object with nine columns, and 150 rows, samples.
}
\source{
Dummy data modified from metagenomic analysis of mouse caecal contents over the day to demonstrate rhythmicity within one group.
}
\usage{
twowaydata
}
\description{
Snippet of example data to demonstrate the functionality of 'kronos' in the two-factor design scenario.
}
\keyword{datasets}
