% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{delistKronos_groupwise}
\alias{delistKronos_groupwise}
\title{Get Results from list of KronosOut Objects}
\usage{
delistKronos_groupwise(kronos_list, padjust = TRUE)
}
\arguments{
\item{kronos_list}{a list of preferrably named kronosOut objects.}

\item{padjust}{a boolean. Toggles FDR using Benjamini Hochbergs procedure.}
}
\value{
A table with circadian output stats per group per feature.
}
\description{
These functions provides a unified wrapper to retrieve results
 from a list of \code{kronosOut} objects.
}
