% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronos.R
\name{fit_cosinor_model}
\alias{fit_cosinor_model}
\title{Fit cosinor model}
\usage{
fit_cosinor_model(
  formula,
  data,
  time = NULL,
  verbose = verbose,
  for_pw = FALSE
)
}
\arguments{
\item{formula}{A formula. Use the \code{time} function to designate which variable represents time.}

\item{data}{Input data}

\item{time}{A string. Should be the column name containing the time values.}

\item{verbose}{A boolean. Toggles whether to print diagnostic information while running. Useful for debugging errors on large datasets.}

\item{for_pw}{A boolean. Toggles whether to perform pairwise ANOVAs as a TukeyHSD-like post-hoc.}
}
\description{
Fit cosinor model for totality of data
}
