% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fit_groupwise_model}
\alias{fit_groupwise_model}
\title{Fit cosinor model}
\usage{
fit_groupwise_model(x, measurement, groups, group, time, period, verbose = T)
}
\description{
Fit cosinor model for one aspect of data
}
