% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronos.R
\name{fit_groupwise_model}
\alias{fit_groupwise_model}
\title{Fit cosinor model}
\usage{
fit_groupwise_model(data, group, time, period, verbose)
}
\arguments{
\item{data}{input data}

\item{group}{A character string. Signifies which group will be assessed.}

\item{time}{A string. Should be the column name containing the time values.}

\item{period}{A numeric. The length of a period, in the same format as the \code{time} parameter.}

\item{verbose}{A boolean. Toggles whether to print diagnostic information while running. Useful for debugging errors on large data sets.}
}
\description{
Fit cosinor model for one aspect of data. Called by main 'kronos' function.
}
