% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getKronos}
\alias{getKronos}
\title{Get Results from KronosOut Object}
\usage{
getKronos(kronosOut, target)
}
\arguments{
\item{kronosOut}{a \code{kronosOut} output object.}

\item{target}{the specific entry of the \code{kronosOut} object to be extracted.}
}
\value{
A \code{data.frame} of results.
}
\description{
These functions provides a unified wrapper to retrieve results
 from a \code{kronosOut} object.
}
