% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{gg_kronos_sinusoid}
\alias{gg_kronos_sinusoid}
\title{A plotting method for circadian plots using ggplot.}
\usage{
gg_kronos_sinusoid(kronosOut, fill = NULL)
}
\description{
Wrapper around ggplot to make figures with a sinusoid trace.
}
