% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronos.R
\name{kronos_predict}
\alias{kronos_predict}
\title{Give tracing information for plotting purposes}
\usage{
kronos_predict(fit, period, time, factors, verbose = verbose)
}
\arguments{
\item{fit}{A model fit}

\item{period}{A numeric. The length of a period, in the same format as the \code{time} parameter.}

\item{time}{A string. Should be the column name containing the time values.}

\item{factors}{A vector. The names of the independent variables.}

\item{verbose}{A boolean. Toggles whether to print diagnostic information while running. Useful for debugging errors on large data sets.}
}
\description{
Generate data needed to plot cosinor trace line.
}
